refactor(asset): F623 — fix cross-domain imports via types.ts (MSA baseline 0)
import AuditBus/EntityRegistry/PolicyEngine from types.ts re-exports
instead of direct service file paths (no-cross-domain-import rule)
